New deployment files to go with repo rename
