+++ Dockerfile-ARM
@@ -1,15 +1,14 @@
-FROM resin/raspberry-pi-alpine-golang
+FROM resin/armv7hf-debian-qemu
 
-RUN apk update && apk add git
-RUN apk add zeromq-dev && apk add build-base
+RUN [ "cross-build-start" ]
 
-RUN mkdir -p $GOPATH/src/github.com/byuoitav/zeromq-proxy-microservice
-ADD . $GOPATH/src/github.com/byuoitav/zeromq-proxy-microservice
+RUN mkdir -p /go
+ADD . /go
 
-WORKDIR $GOPATH/src/github.com/byuoitav/zeromq-proxy-microservice
-RUN go get -d -v
-RUN go install -v
+WORKDIR /go
 
-CMD ["/go/bin/zeromq-proxy-microservice"]
+CMD ["/go/event-router-microservice-arm"]
 
 EXPOSE 7000
+
+RUN [ "cross-build-end" ]
