+++ Dockerfile-branch
@@ -3,18 +3,14 @@ FROM golang:alpine
 ARG GITHUB_BRANCH
 
 RUN apk update && apk add git
-RUN apk add zeromq
-RUN go get -u github.com/FiloSottile/gvt
 
 RUN mkdir -p /go/src/github.com/byuoitav
-ADD . /go/src/github.com/byuoitav/zeromq-proxy-microservice
+ADD . /go/src/github.com/byuoitav/event-router-microservice
 
-WORKDIR /go/src/github.com/byuoitav/zeromq-proxy-microservice
-RUN gvt fetch -branch ${GITHUB_BRANCH} github.com/byuoitav/configuration-database-microservice
-RUN gvt fetch -branch ${GITHUB_BRANCH} github.com/byuoitav/av-api
+WORKDIR /go/src/github.com/byuoitav/event-router-microservice
 RUN go get -d -v
 RUN go install -v
 
-CMD ["/go/bin/zeromq-proxy-microservice"]
+CMD ["/go/bin/event-router-microservice"]
 
-EXPOSE 7000
+EXPOSE 8004
