+++ circle.yml
@@ -1,6 +1,4 @@
 machine:
-    pre:
-        - sudo curl -sSL https://s3.amazonaws.com/circle-downloads/install-circleci-docker.sh | bash -s -- 1.10.0
     services:
         - docker
     environment:
@@ -12,22 +10,11 @@ machine:
 dependencies:
     pre:
         - rm -rf ~/.go_workspace
-        - go get -u github.com/FiloSottile/gvt
-        - sudo apt update && sudo apt install libzmq3-dev
-        - sudo docker run --privileged multiarch/qemu-user-static:register --reset
     override:
         - mkdir -p ~/.go_project/src/github.com/${CIRCLE_PROJECT_USERNAME}
         - ln -s ${HOME}/${CIRCLE_PROJECT_REPONAME} ${HOME}/.go_project/src/github.com/${CIRCLE_PROJECT_USERNAME}/${CIRCLE_PROJECT_REPONAME}
         - cd ~/.go_project/src/github.com/${CIRCLE_PROJECT_USERNAME}/${CIRCLE_PROJECT_REPONAME}
         - go get -d -v
-        - >
-            if [ $CIRCLE_BRANCH = "stage" ]; then
-                gvt fetch -branch stage github.com/byuoitav/configuration-database-microservice
-                gvt fetch -branch stage github.com/byuoitav/av-api
-            elif [ $CIRCLE_BRANCH = "production" ]; then
-                gvt fetch -branch production github.com/byuoitav/configuration-database-microservice
-                gvt fetch -branch production github.com/byuoitav/av-api
-            fi
 
 compile:
     override:
@@ -41,7 +28,7 @@ deployment:
     development:
         branch: master
         commands:
-            - docker run --privileged multiarch/qemu-user-static:register --reset
+            - cd $BUILD_PATH && env GOOS=linux GOARCH=arm go build -v -o ${CIRCLE_PROJECT_REPONAME}-arm
             - docker build -f Dockerfile-ARM -t byuoitav/rpi-${CIRCLE_PROJECT_REPONAME}:development .
             - docker build -f Dockerfile-development -t byuoitav/${CIRCLE_PROJECT_REPONAME}:development .
             - docker login -e $DOCKER_EMAIL -u $DOCKER_USERNAME -p $DOCKER_PASSWORD
@@ -51,7 +38,7 @@ deployment:
     stage:
         branch: stage
         commands:
-            - docker run --privileged multiarch/qemu-user-static:register --reset
+            - cd $BUILD_PATH && env GOOS=linux GOARCH=arm go build -v -o ${CIRCLE_PROJECT_REPONAME}-arm
             - docker build -f Dockerfile-ARM -t byuoitav/rpi-${CIRCLE_PROJECT_REPONAME}:stage .
             - docker build --build-arg GITHUB_BRANCH=stage -f Dockerfile-branch -t byuoitav/${CIRCLE_PROJECT_REPONAME}:stage .
             - docker login -e $DOCKER_EMAIL -u $DOCKER_USERNAME -p $DOCKER_PASSWORD
@@ -61,10 +48,11 @@ deployment:
     production:
         branch: production
         commands:
-            - docker run --privileged multiarch/qemu-user-static:register --reset
+            - cd $BUILD_PATH && env GOOS=linux GOARCH=arm go build -v -o ${CIRCLE_PROJECT_REPONAME}-arm
             - docker build -f Dockerfile-ARM -t byuoitav/rpi-${CIRCLE_PROJECT_REPONAME}:latest .
             - docker build --build-arg GITHUB_BRANCH=production -f Dockerfile-branch -t byuoitav/${CIRCLE_PROJECT_REPONAME}:latest .
             - docker login -e $DOCKER_EMAIL -u $DOCKER_USERNAME -p $DOCKER_PASSWORD
             - docker push byuoitav/${CIRCLE_PROJECT_REPONAME}:latest
             - docker push byuoitav/rpi-${CIRCLE_PROJECT_REPONAME}:latest
+            - ./deploy.sh $CIRCLE_PROJECT_REPONAME $CIRCLE_SHA1
             - ./deploy-pi.sh production
