+++ server.go
@@ -7,8 +7,7 @@ import (
 	"sync"
 
 	"github.com/byuoitav/av-api/dbo"
-	"github.com/byuoitav/zeromq-proxy-miroservice/proxy"
-	"github.com/byuoitav/zeromq-proxy-miroservice/zeromq"
+	"github.com/byuoitav/event-router-microservice/tags"
 )
 
 func main() {
@@ -28,19 +27,21 @@ func main() {
 
 	addresses := []string{}
 	for _, device := range devices {
-		addresses = append(addresses, device.Address)
+		addresses = append(addresses, device.Address+":7000")
 	}
 
-	inChan := make(chan zeromq.Event, 100)
-	outChan := make(chan zeromq.Event, 100)
-	exitChan := make(chan bool, 3)
+	//subscribe to the av-api and the event translator on the local pi
+	addresses = append(addresses, "localhost:7001", "localhost:7002")
 
-	log.Printf("Starting publisher")
-	go proxy.Publish(outChan, exitChan, port, wg)
-	log.Printf("Starting router")
-	go proxy.Router(inChan, outChan, exitChan, wg)
-	log.Printf("Starting Receiver")
-	go proxy.Recieve(inChan, exitChan, addresses, wg)
+	RoutingTable := make(map[string][]string)
+	RoutingTable[tags.LocalAPI] = []string{tags.TransmitAPI}
+	RoutingTable[tags.TransmitAPI] = []string{tags.LocalTransmit}
+	RoutingTable[tags.External] = []string{tags.LocalTransmit}
+
+	err := router.Start(RoutingTable, wg, 1000, addresses, 120, "7000")
+	if err != nil {
+		fmt.Error(err)
+	}
 
 	wg.Wait()
 }
